fix: move ApiUserSchema to types.py
